practice(ReviewResponseDTO): 생성한 DTO response로 이동
